refactor!: simplify exception namespace
