+++ tests/features/steps/steps.py
@@ -3,7 +3,7 @@
 from behave import given, then, when
 
 from open_feature.evaluation_context import EvaluationContext
-from open_feature.exception.error_code import ErrorCode
+from open_feature.exception import ErrorCode
 from open_feature.flag_evaluation.flag_evaluation_details import FlagEvaluationDetails
 from open_feature.flag_evaluation.reason import Reason
 from open_feature.open_feature_api import get_client, set_provider
