+++ tests/test_open_feature_api.py
@@ -4,8 +4,7 @@ import pytest
 
 from open_feature.evaluation_context import EvaluationContext
 from open_feature.hooks.hook import Hook
-from open_feature.exception.error_code import ErrorCode
-from open_feature.exception.exceptions import GeneralError
+from open_feature.exception import ErrorCode, GeneralError
 from open_feature.open_feature_api import (
     get_client,
     get_provider,
