+++ tests/test_open_feature_client.py
@@ -3,8 +3,7 @@ from unittest.mock import MagicMock
 import pytest
 
 from open_feature.open_feature_api import add_hooks, clear_hooks
-from open_feature.exception.error_code import ErrorCode
-from open_feature.exception.exceptions import OpenFeatureError
+from open_feature.exception import ErrorCode, OpenFeatureError
 from open_feature.flag_evaluation.reason import Reason
 from open_feature.hooks.hook import Hook
 from open_feature.open_feature_client import OpenFeatureClient
